+++ examples/data-sources/pingfederate_oauth_token_exchange_token_generator_mapping/data-source.tf
@@ -1,0 +1,3 @@
+data "pingfederate_oauth_token_exchange_token_generator_mapping" "oauthTokenExchangeTokenGeneratorMappingsExample" {
+  mapping_id = "example"
+}