+++ internal/acctest/keypairsigningimport/key_pair_signing_import_test.go
@@ -81,7 +81,7 @@ resource "pingfederate_key_pair_signing_import" "%[1]s" {
 }
 
 data "pingfederate_key_pair_signing_import" "%[1]s" {
-  id = pingfederate_key_pair_signing_import.%[1]s.id
+  import_id = pingfederate_key_pair_signing_import.%[1]s.id
 }`, resourceName,
 		resourceModel.id,
 		resourceModel.fileData,