+++ internal/acctest/tokenprocessortotokengeneratormapping/token_processor_to_token_generator_mapping_test.go
@@ -100,7 +100,7 @@ resource "pingfederate_token_processor_to_token_generator_mapping" "%[1]s" {
 	%[7]s
 }
 data "pingfederate_token_processor_to_token_generator_mapping" "%[1]s" {
-  id = pingfederate_token_processor_to_token_generator_mapping.%[1]s.id
+  mapping_id = pingfederate_token_processor_to_token_generator_mapping.%[1]s.id
 }`, resourceName,
 		resourceModel.sourceId,
 		resourceModel.targetId,