+++ internal/resource/config/authenticationpolicycontract/authentication_policy_contract_data_source.go
@@ -67,11 +67,10 @@ func (r *authenticationPolicyContractDataSource) Schema(ctx context.Context, req
 		},
 	}
 
-	id.ToSchema(&schema)
+	id.ToDataSourceSchema(&schema)
 	id.ToDataSourceSchemaCustomId(&schema,
 		"contract_id",
 		true,
-		false,
 		"The persistent, unique ID for the authentication policy contract.")
 	resp.Schema = schema
 }