+++ internal/resource/config/oauth/authserversettings/scopes/commonscope/oauth_auth_server_settings_scopes_common_scope_data_source.go
@@ -59,7 +59,7 @@ func (r *oauthAuthServerSettingsScopesCommonScopeDataSource) Schema(ctx context.
 			},
 		},
 	}
-	id.ToDataSourceSchema(&schemaDef, false, "The ID of this resource.")
+	id.ToDataSourceSchema(&schemaDef)
 	resp.Schema = schemaDef
 }