+++ internal/resource/config/serversettings/generalsettings/server_settings_general_settings_data_source.go
@@ -71,7 +71,7 @@ func (r *serverSettingsGeneralSettingsDataSource) Schema(ctx context.Context, re
 		},
 	}
 
-	id.ToSchema(&schema)
+	id.ToDataSourceSchema(&schema)
 	resp.Schema = schema
 }