+++ internal/resource/config/serversettings/logsettings/server_settings_log_settings_data_source.go
@@ -72,7 +72,7 @@ func (r *serverSettingsLogSettingsDataSource) Schema(ctx context.Context, req da
 		},
 	}
 
-	id.ToSchema(&schema)
+	id.ToDataSourceSchema(&schema)
 	resp.Schema = schema
 }