+++ internal/resource/config/tokenprocessortotokengeneratormapping/token_processor_to_token_generator_mapping_common.go
@@ -17,6 +17,7 @@ type tokenProcessorToTokenGeneratorMappingModel struct {
 	IssuanceCriteria                 types.Object `tfsdk:"issuance_criteria"`
 	SourceId                         types.String `tfsdk:"source_id"`
 	TargetId                         types.String `tfsdk:"target_id"`
+	MappingId                        types.String `tfsdk:"mapping_id"`
 	Id                               types.String `tfsdk:"id"`
 	DefaultTargetResource            types.String `tfsdk:"default_target_resource"`
 	LicenseConnectionGroupAssignment types.String `tfsdk:"license_connection_group_assignment"`
@@ -32,6 +33,7 @@ func readTokenProcessorToTokenGeneratorMappingResponse(ctx context.Context, r *c
 	diags.Append(respDiags...)
 	state.SourceId = types.StringValue(r.SourceId)
 	state.TargetId = types.StringValue(r.TargetId)
+	state.MappingId = types.StringPointerValue(r.Id)
 	state.Id = types.StringPointerValue(r.Id)
 	state.DefaultTargetResource = types.StringPointerValue(r.DefaultTargetResource)
 	state.LicenseConnectionGroupAssignment = types.StringPointerValue(r.LicenseConnectionGroupAssignment)