+++ internal/resource/config/tokenprocessortotokengeneratormapping/token_processor_to_token_generator_mapping_data_source.go
@@ -61,7 +61,8 @@ func (r *tokenProcessorToTokenGeneratorMappingDataSource) Schema(ctx context.Con
 			"issuance_criteria": issuancecriteria.ToDataSourceSchema(),
 		},
 	}
-	id.ToDataSourceSchema(&schema, true, "ID of Token Processor to Token Generator Mapping.")
+	id.ToDataSourceSchema(&schema)
+	id.ToDataSourceSchemaCustomId(&schema, "mapping_id", true, "ID of Token Processor to Token Generator Mapping.")
 	resp.Schema = schema
 }
 
@@ -89,7 +90,7 @@ func (r *tokenProcessorToTokenGeneratorMappingDataSource) Read(ctx context.Conte
 	if resp.Diagnostics.HasError() {
 		return
 	}
-	apiReadTokenProcessorToTokenGeneratorMapping, httpResp, err := r.apiClient.TokenProcessorToTokenGeneratorMappingsAPI.GetTokenToTokenMappingById(config.ProviderBasicAuthContext(ctx, r.providerConfig), state.Id.ValueString()).Execute()
+	apiReadTokenProcessorToTokenGeneratorMapping, httpResp, err := r.apiClient.TokenProcessorToTokenGeneratorMappingsAPI.GetTokenToTokenMappingById(config.ProviderBasicAuthContext(ctx, r.providerConfig), state.MappingId.ValueString()).Execute()
 
 	if err != nil {
 		config.ReportHttpError(ctx, &resp.Diagnostics, "An error occurred while getting a Token Processor To Token Generator Mapping", err, httpResp)