+++ internal/resource/config/virtualhostnames/virtual_host_names_data_source.go
@@ -36,7 +36,7 @@ type virtualHostNamesDataSourceModel struct {
 
 // GetSchema defines the schema for the datasource.
 func (r *virtualHostNamesDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
-	schemaDef := schema.Schema{
+	schema := schema.Schema{
 		Description: "Describes settings for virtual host names.",
 		Attributes: map[string]schema.Attribute{
 			"virtual_host_names": schema.ListAttribute{
@@ -47,8 +47,8 @@ func (r *virtualHostNamesDataSource) Schema(ctx context.Context, req datasource.
 			},
 		},
 	}
-	id.ToSchema(&schemaDef)
-	resp.Schema = schemaDef
+	id.ToDataSourceSchema(&schema)
+	resp.Schema = schema
 }
 
 // Metadata returns the data source type name.